SWITCHYARD-1536 eliminate split packages in core
