+++ transform/src/main/java/org/switchyard/transform/config/model/v1/V1JAXBTransformModel.java
@@ -25,8 +25,8 @@ import org.switchyard.config.Configuration;
 import org.switchyard.config.model.Descriptor;
 import org.switchyard.config.model.transform.TransformModel;
 import org.switchyard.config.model.transform.v1.V1BaseTransformModel;
-import org.switchyard.transform.TransformerFactoryClass;
 import org.switchyard.transform.config.model.JAXBTransformModel;
+import org.switchyard.transform.internal.TransformerFactoryClass;
 import org.switchyard.transform.jaxb.internal.JAXBTransformerFactory;
 
 /**
