+++ transform/src/main/java/org/switchyard/transform/config/model/v1/V1JSONTransformModel.java
@@ -21,12 +21,12 @@ package org.switchyard.transform.config.model.v1;
 
 import javax.xml.namespace.QName;
 
-import org.switchyard.transform.TransformerFactoryClass;
 import org.switchyard.config.Configuration;
 import org.switchyard.config.model.Descriptor;
 import org.switchyard.config.model.transform.TransformModel;
 import org.switchyard.config.model.transform.v1.V1BaseTransformModel;
 import org.switchyard.transform.config.model.JSONTransformModel;
+import org.switchyard.transform.internal.TransformerFactoryClass;
 import org.switchyard.transform.json.internal.JSONTransformFactory;
 
 /**
