+++ transform/src/main/java/org/switchyard/transform/config/model/v1/V1SmooksTransformModel.java
@@ -21,13 +21,13 @@ package org.switchyard.transform.config.model.v1;
 
 import javax.xml.namespace.QName;
 
-import org.switchyard.transform.TransformerFactoryClass;
 import org.switchyard.config.Configuration;
 import org.switchyard.config.model.Descriptor;
 import org.switchyard.config.model.transform.TransformModel;
 import org.switchyard.config.model.transform.v1.V1BaseTransformModel;
 import org.switchyard.transform.config.model.SmooksTransformModel;
 import org.switchyard.transform.config.model.SmooksTransformType;
+import org.switchyard.transform.internal.TransformerFactoryClass;
 import org.switchyard.transform.smooks.internal.SmooksTransformFactory;
 
 /**
