+++ transform/src/main/java/org/switchyard/transform/xslt/internal/XsltTransformFactory.java
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  * MA  02110-1301, USA.
  */
-package org.switchyard.transform.xslt;
+package org.switchyard.transform.xslt.internal;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -32,8 +32,8 @@ import org.apache.log4j.Logger;
 import org.switchyard.common.type.Classes;
 import org.switchyard.exception.SwitchYardException;
 import org.switchyard.transform.Transformer;
-import org.switchyard.transform.TransformerFactory;
 import org.switchyard.transform.config.model.XsltTransformModel;
+import org.switchyard.transform.internal.TransformerFactory;
 
 /**
  * @author Alejandro Montenegro <a href="mailto:aamonten@gmail.com">aamonten@gmail.com</a>
