+++ transform/src/test/java/org/switchyard/transform/AbstractTransformerTestCase.java
@@ -29,6 +29,7 @@ import org.switchyard.config.model.switchyard.SwitchYardModel;
 import org.switchyard.config.model.transform.TransformModel;
 import org.switchyard.config.model.transform.TransformsModel;
 import org.switchyard.internal.transform.BaseTransformerRegistry;
+import org.switchyard.transform.internal.TransformerRegistryLoader;
 
 /**
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
