+++ transform/src/test/java/org/switchyard/transform/internal/json/JSONTransformerTest.java
@@ -41,8 +41,8 @@ import org.switchyard.transform.Message2MessageTransformer;
 import org.switchyard.transform.TransformSequence;
 import org.switchyard.transform.Transformer;
 import org.switchyard.transform.TransformerRegistry;
-import org.switchyard.transform.TransformerRegistryLoader;
 import org.switchyard.transform.config.model.JSONTransformModel;
+import org.switchyard.transform.internal.TransformerRegistryLoader;
 import org.switchyard.transform.internal.json.User.Gender;
 import org.switchyard.transform.internal.json.User.Name;
 import org.switchyard.transform.json.internal.JSON2JavaTransformer;
